WIP: more progress on contact authors workflow
Current workflow has curator select a status (code not available, code
has a URL but not one we consider "stable", code is available), the
number of publications, and some optional custom invitation text, then
clicks "Submit" to see a preview of the email.

If they are happy with the way the email looks, check the Ready to send
button and click Submit and it'll send the email out and redirect
elsewhere.

Still unfinished, need to:

- wire up the preview email text to the page properly
- adjust email templates to have proper default values for empty strings
when in preview mode
- set up email filter